Add basic support for Microsoft Graph
