+++ BlazorForms.Core/BlazorForms.Core.csproj
@@ -15,6 +15,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Azure.Identity" Version="1.21.0" />
 		<PackageReference Include="DbController" Version="5.0.0" />
 		<PackageReference Include="DbController.MySql" Version="5.0.0" />
 		<PackageReference Include="DbController.TypeHandler" Version="5.0.0" />
@@ -31,6 +32,7 @@
 		<PackageReference Include="Microsoft.Extensions.Hosting" Version="10.0.8" />
 		<PackageReference Include="Microsoft.Extensions.Identity.Core" Version="10.0.8" />
 		<PackageReference Include="Microsoft.Extensions.Localization.Abstractions" Version="10.0.8" />
+		<PackageReference Include="Microsoft.Graph" Version="6.1.0" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.4" />
 	</ItemGroup>
