+++ BlazorForms.Core/Infrastructure/Email/EmailMessage.cs
@@ -1,0 +1,12 @@
+﻿namespace BlazorForms.Core.Infrastructure;
+
+public class EmailMessage
+{
+    public List<string> To { get; set; } = [];
+    public List<string> Cc { get; set; } = [];
+    public List<string> Bcc { get; set; } = [];
+    public List<(string filename, byte[] content)> Attachments { get; set; } = [];
+    public string Subject { get; set; } = string.Empty;
+    public string Body { get; set; } = string.Empty;
+
+}
