+++ BlazorForms.Core/Infrastructure/Email/IEmailSender.cs
@@ -1,0 +1,8 @@
+﻿namespace BlazorForms.Core.Infrastructure;
+
+public interface IEmailSender
+{
+    MailProvider Provider { get; }
+
+    Task SendAsync(EmailMessage message);
+}
