+++ BlazorForms.Core/Infrastructure/Email/IEmailSenderFactory.cs
@@ -1,0 +1,6 @@
+﻿namespace BlazorForms.Core.Infrastructure;
+
+public interface IEmailSenderFactory
+{
+    IEmailSender CreateSender(MailProvider provider);
+}
