+++ BlazorForms.Core/Infrastructure/Email/MailProvider.cs
@@ -1,0 +1,8 @@
+﻿namespace BlazorForms.Core.Infrastructure;
+
+public enum MailProvider
+{
+    None = 0,
+    Smtp = 1,
+    MicrosoftGraph = 2
+}
