+++ BlazorForms.Core/Infrastructure/Email/NoneSender.cs
@@ -1,0 +1,11 @@
+﻿namespace BlazorForms.Core.Infrastructure;
+
+public class NoneSender : IEmailSender
+{
+    public MailProvider Provider => MailProvider.None;
+
+    public Task SendAsync(EmailMessage message)
+    {
+        return Task.CompletedTask;
+    }
+}
