+++ BlazorForms/Components/Pages/Admin/Forms/FormEditor.razor
@@ -4,7 +4,6 @@
 @using BlazorForms.Components.Admin.Forms
 @layout EmptyLayout
 @inherits BlazorFormsComponentBase
-@inject IOptions<EmailSettings> emailConfig
 @inject FormService formService
 @inject UserService userService
 @inject NavigationManager navigationManager
