+++ BlazorForms/Components/_Imports.razor
@@ -20,6 +20,7 @@
 @using BlazorForms.Core.Interfaces
 @using BlazorForms.Core.Models
 @using BlazorForms.Core.Filters
+@using BlazorForms.Core.Infrastructure
 @using BlazorForms.Core.Models.FormElements
 @using BlazorForms.Core.Extensions
 @using CKEditor
@@ -28,7 +29,6 @@
 @using System.Globalization
 @using Microsoft.Extensions.Localization
 @using Microsoft.Extensions.Options
-@using BlazorForms.Core.Settings
 @using BlazorBootstrap
 @using BlazorForms.Components.Layout
 @using Toolbelt.Blazor.HotKeys2
